requirements.txt
langchain==0.1.17
google-generativeai==0.4.1
pandas==2.2.2
streamlit==1.35.0
python-dotenv
langchain-google-genai


#for code execution after user request:

matplotlib
seaborn
numpy

